oci: Handle references of images not yet pushed
This commit adds a change to the OCI library to generate
identifiers for images which do not exist in the registry.

Whe we have enough data in the string we use it to form the
identifiers. If the lookup for variants fails, we still
return what we know from the identifier string.

Signed-off-by: Adolfo Garcia Veytia (puerco) <puerco@chainguard.dev>